now builds complete SDKs for both x86 and a64

+++ _builtins.sh
@@ -538,4 +538,4 @@ for TARGET_TRIPLE in ${TARGET_TRIPLES[@]}; do
 done
 
 _popd
-rm -rf "$BUILDDIR"
+$NO_CLEANUP || rm -rf "$BUILDDIR"
